+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * **Fix**: [423](https://github.com/SimformSolutionsPvtLtd/chatview/pull/423)
   Rendering issue in attached image preview when sending message on web.
+* **Feat**: [420](https://github.com/SimformSolutionsPvtLtd/chatview/pull/420) Added support for
+  `playerMode` in `VoiceMessageConfiguration` with `single` and `multi`.
 
 ## [3.0.0]